+++ backend/app/ad.py
@@ -1,0 +1,4 @@
+# import os
+# current_path = os.path.abspath(__file__)
+# root_folder = os.path.dirname(current_path)
+# root_folder = str.replace(root_folder, "app", "")
