+++ backend/app/models/__init__.py
@@ -1,0 +1,1 @@
+print(__name__, "MODEL")
