+++ backend/app/models/storage_engine/__init__.py
@@ -1,6 +1,6 @@
 from models.storage_engine.connect_db import ConnectMysqlDB  # type: ignore
 from models.storage_engine.db import DBStorage  # type: ignore
-
+storage = None
 if __name__ == "models.storage_engine":
     """Instantiate and initialize the DBStorage if connection was successful"""
     db_connection = ConnectMysqlDB()
