+++ backend/app/resouces/__init__.py
@@ -3,13 +3,14 @@
 from flask import Blueprint  # type: ignore
 from flask_restful import Api  # type: ignore
 
-from .expenses import ExpenseResource  # type: ignore
-from .product import ProductResource  # type: ignore
-from .revenue import RevenueResource  # type: ignore
-
 resource_product_revenue_route = Blueprint(
     "route_resource_product_revenue", __name__, url_prefix="/api/v1/"
 )
+from resouces.expenses import ExpenseResource  # type: ignore
+from resouces.product import ProductResource  # type: ignore
+from resouces.revenue import RevenueResource  # type: ignore
+
+
 # import the token_route  which is the blueprint
 resource_product_revenue_route_api = Api(resource_product_revenue_route)
