+++ backend/run_test.sh
@@ -6,5 +6,6 @@ set -eou pipefail
 script_path=$(readlink -f "$0")
 FILE_PATH=${script_path%/*}
 
-set APP_ENV=test && set MYSQL_USER=localhost && "C:\Users\Adeyori\alx-project\backend\alx-venv\Scripts\python.exe" -m pytest "${FILE_PATH}\tests"
+# set APP_ENV=test && set MYSQL_USER=localhost && "${FILE_PATH}/alx_venv/Scripts/python.exe" -m pytest "${FILE_PATH}/tests"
 
+ python -m unittest discover -s tests
