+++ backend/tests/__init__.py
@@ -1,0 +1,3 @@
+import sys
+print(sys.path)
+print(__name__, "TEST")
